feat(query): add ownership check in all queries

+++ internal/domain/repository/project/project.sql
@@ -1,7 +1,9 @@
 -- name: GetAllProjects :many
 SELECT
   *
-FROM public.project p;
+FROM public.project p
+WHERE
+  p.user_id = $1 ;
 
 -- name: GetProjectById :one
 SELECT
@@ -9,19 +11,24 @@ SELECT
 FROM
   public.project p
 WHERE
-  p.id = $1;
+  p.id = $1
+  AND p.user_id = $2;
 
 -- name: CreateProject :exec
-INSERT INTO public.project (name,description)
-VALUES ($1,$2);
+INSERT INTO public.project (user_id,name,description)
+VALUES ($1,$2,$3);
 
 -- name: UpdateProjectById :exec
 UPDATE public.project
 SET name = $1,
     description = $2,
     updated_at = now()
-WHERE id = $3;
+WHERE
+  id = $3
+  AND user_id = $4;
 
 -- name: DeleteProjectById :exec
 DELETE FROM public.project
-where id = $1;
+WHERE
+  id = $1
+  AND user_id = $2;
